i2c: i2c-algo-bit: Add support for bus recovery
If I2C master stops in the middle of I2C transaction, to a
slave, the it may leave the slave device in an incomplete
state. Typical case is master fails when doing a read operation
in a slave. SDA line is non usable anymore because of the
"Slave-Transmitter" mode.

Methods to recover the SDA line:
 1. Reset the slave device (if it has Reset pin)
 2. Use bus recovery sequence.

Bus recovery sequence steps:
 1. Master tries to pull up SDA line
 2. Master generates a clock pulse on SCL (1-0-1 transition)
 3. Master checks SDA. If SDA = 0 goto Step 2 until 9 iterations
    completed; if SDA = 1 goto Step 4.
 4. Master sends a STOP command initializing completely the bus.

Add recovery support for i2c-gpio bit bang driver during
registration of adapter. Call the bus recovery if SDA is
low before starting a new i2c transaction.

CRs-Fixed: 468622
Change-Id: I115a1bc20defbdf6412fd0eb5d4436f550136a0f
Signed-off-by: Tirupathi Reddy <tirupath@codeaurora.org>